Move helper functions to internal
